refactor JS
