+++ jdi-nova/src/main/java/com/jdiai/CoreElement.java
@@ -3,8 +3,12 @@ package com.jdiai;
 import com.jdiai.interfaces.HasCore;
 
 public class CoreElement implements HasCore {
-    private JS core = new JS(JDI::driver);
-    public JS core() { return core; }
+    private JS core;
+
+    public JS core() {
+        return core != null ? core : JDI.initJSFunc.get();
+    }
+
     public void setCore(JS core) { this.core = core; }
 
 }
