+++ jdi-nova/src/main/java/com/jdiai/Section.java
@@ -1,9 +1,3 @@
 package com.jdiai;
 
-import com.jdiai.interfaces.HasCore;
-
-public class Section implements HasCore {
-    JS core = new JS(JDI::driver);
-    public JS core() { return core; }
-    public void setCore(JS core) { this.core = core; }
-}
+public class Section extends CoreElement { }
