+++ jdi-nova/src/main/java/com/jdiai/interfaces/HasName.java
@@ -3,4 +3,5 @@ package com.jdiai.interfaces;
 public interface HasName {
     String getName();
     <T> T setName(String name);
+
 }
