+++ jdi-nova/src/main/java/com/jdiai/page/objects/PageFactoryUtils.java
@@ -5,14 +5,12 @@ import com.jdiai.JS;
 import com.jdiai.annotations.Title;
 import com.jdiai.annotations.Url;
 import com.jdiai.interfaces.HasCore;
-import com.jdiai.interfaces.HasLocators;
 import com.jdiai.jsdriver.JDINovaException;
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
 
 import java.lang.reflect.Constructor;
 import java.lang.reflect.Field;
-import java.util.ArrayList;
 import java.util.List;
 
 import static com.epam.jdi.tools.LinqUtils.*;
@@ -20,6 +18,7 @@ import static com.epam.jdi.tools.ReflectionUtils.getFieldsDeep;
 import static com.epam.jdi.tools.ReflectionUtils.isInterface;
 import static com.jdiai.page.objects.JDIPageFactory.LOCATOR_FROM_FIELD;
 import static com.jdiai.page.objects.PageFactory.getFactory;
+import static com.jdiai.tools.JSUtils.getLocators;
 import static java.lang.String.format;
 import static org.apache.commons.lang3.ObjectUtils.isEmpty;
 
@@ -50,21 +49,16 @@ public class PageFactoryUtils {
     static void setupCoreElement(InitInfo info) {
         By locator = LOCATOR_FROM_FIELD.apply(info.field);
         JS core;
+        List<By> locators = getLocators(info.parent);
+        core = JDI.initJSFunc.get();
         if (locator != null) {
-            core = new JS(JDI::driver, locator, info.parent);
-        } else {
-            List<By> locators = new ArrayList<>();
-            if (info.parent != null && isInterface(info.parent.getClass(), HasLocators.class)) {
-                locators = ((HasLocators) info.parent).locators();
-            }
-            core = isEmpty(locators)
-                ? new JS()
-                : new JS(JDI::driver, locators);
+            locators.add(locator);
         }
+        core.setLocators(locators);
         ((HasCore) info.instance).setCore(core);
     }
     static boolean isUIObject(Field field) {
-        if (field.getName().equals("core") || field.getType().isAssignableFrom(JS.class)) {
+        if (field.getName().equals("core") || isInterface(field.getType(), JS.class)) {
             return false;
         }
         List<Field> fields = getFieldsDeep(field);
