+++ jdi-nova/src/main/java/com/jdiai/tools/FilterConditions.java
@@ -8,11 +8,11 @@ import static java.lang.String.format;
 
 public class FilterConditions {
     public static Function<JS, String> textEquals(String value) {
-        return el -> format("%s === '%s'", el.textType.value.trim(), value.trim());
+        return el -> format("%s === '%s'", el.textType().trim(), value.trim());
     }
     public static Function<JS, String> textContains(String value) {
         return el -> format("%s.toLowerCase().includes('%s'.toLowerCase())",
-            el.textType.value, value);
+            el.textType(), value);
     }
     public static Function<JS, String> hasAttribute(String attrName) {
         return el -> "hasAttribute('" + attrName + "')";
