+++ jdi-nova/src/main/java/com/jdiai/tools/VisualSettings.java
@@ -11,7 +11,7 @@ import java.util.function.BiFunction;
 import java.util.function.Function;
 
 import static com.epam.jdi.tools.ReflectionUtils.isInterface;
-import static com.jdiai.JS.JDI_STORAGE;
+import static com.jdiai.JDI.JDI_STORAGE;
 import static com.jdiai.visual.ImageTypes.JPG;
 import static java.lang.Math.*;
 import static java.lang.String.format;
